backend/requirements.txt
fastapi
uvicorn[standard]
httpx
google-generativeai
apscheduler
africastalking
python-dotenv
aiosqlite
pydantic
